setup base full
