+++ sep490_mobile/lib/exceptions/custom_exception.dart
@@ -3,11 +3,13 @@ class CustomException implements Exception {
   final String message;
   static const int INTERNAL_SERVER_ERROR = 500;
   static const int UNAUTHORIZED = 401;
-  CustomException({
-    required this.statusCode,
-    required this.message
-  });
 
+  /// Creates a custom exception with the specified status code and message.
+  ///
+  /// Both [statusCode] and [message] parameters are required.
+  CustomException({required this.statusCode, required this.message});
+
+  /// Returns a string representation of the exception.
   @override
   String toString() {
     return 'MyCustomException: Status Code $statusCode, Message: $message';
