--- sep490_mobile/lib/extensions/custon_string.dart
@@ -1,6 +1,0 @@
-import 'dart:convert' as convert;
-extension CustomStringExtensions on String {
-  String toUtf8() {
-    return convert.utf8.decode(this.codeUnits);
-  }
-}
