+++ sep490_mobile/lib/widgets/loading.dart
@@ -7,14 +7,24 @@ class Loading extends StatelessWidget {
   final Color leftDotColor;
   final Color rightDotColor;
 
-  const Loading({super.key,
+  /// A loading animation widget that displays twisting dots.
+  ///
+  /// Parameters:
+  /// - [size]: The size of the loading animation (default: 50)
+  /// - [leftDotColor]: Color of the left dot (default: AppColors.primaryColor)
+  /// - [rightDotColor]: Color of the right dot (default: Colors.blue)
+
+  const Loading({
+    super.key,
     this.size = 100, // Set default value for size
-    this.leftDotColor = AppColors.primaryColor, // Set default value for leftDotColor
+    this.leftDotColor =
+        AppColors.primaryColor, // Set default value for leftDotColor
     this.rightDotColor = Colors.blue, // Set default value for rightDotColor
   });
 
   @override
   Widget build(BuildContext context) {
+    assert(size > 0, 'Size must be positive');
     return Center(
       child: LoadingAnimationWidget.twistingDots(
         leftDotColor: leftDotColor,
